+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix `HeaderNav` layout.
+
 ## [2.66.0] - 2020-04-24
 
 Features: