+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update `path` fill on `Button`.
+
 ## [2.66.0] - 2020-04-24
 
 Features: