+++ CHANGELOG.md
@@ -4,8 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Fix `HeaderNav` layout.
-- Fix: Update `path` fill on `Button`.
+## [2.66.1] - 2020-04-27
+
+Fixes:
+- Fix `HeaderNav` layout.
+- Update `path` fill on `Button`.
 
 ## [2.66.0] - 2020-04-24